@@ -19,28 +19,27 @@
     ]
   },
   "dependencies": {
-    "@next/font": "^13.1.6",
-    "antd": "^5.2.1",
-    "axios": "^1.3.4",
+    "@ant-design/cssinjs": "^1.17.0",
+    "antd": "^5.8.5",
+    "axios": "^1.5.0",
     "date-fns": "^2.29.3",
     "jsonwebtoken": "^9.0.0",
-    "next": "13.1.6",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "styled-components": "^5.3.6",
-    "yup": "^1.0.0"
+    "next": "^13.4.19",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "styled-components": "^6.0.7",
+    "yup": "^1.2.0"
   },
   "devDependencies": {
     "@types/jsonwebtoken": "^9.0.2",
     "@types/node": "18.13.0",
-    "@types/react": "18.0.28",
+    "@types/react": "^18.2.21",
     "@types/react-dom": "18.0.11",
-    "@types/styled-components": "^5.1.26",
     "@typescript-eslint/eslint-plugin": "^5.52.0",
     "eslint": "^8.34.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-config-next": "13.1.6",
+    "eslint-config-next": "^13.4.19",
     "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-jsx-a11y": "^6.7.1",
@@ -50,6 +49,6 @@
     "husky": "^8.0.0",
     "lint-staged": "^13.1.2",
     "prettier": "^2.8.4",
-    "typescript": "4.9.5"
+    "typescript": "^5.2.2"
   }
 }
