@@ -1,7 +1,7 @@
-import { FC } from '@/types';
-import { FormItemProps as AntdFormItemProps } from 'antd';
+import { ComponentProps, FC } from '@/types';
+import { Form } from 'antd';
 import { Schema } from 'yup';
 
-type FormItemProps = { schema?: Schema } & AntdFormItemProps;
+type FormItemProps = { schema?: Schema } & ComponentProps<typeof Form.Item>;
 
 export type FormItemFC = FC<FormItemProps>;
