@@ -1,5 +1,5 @@
-import { PropsWithChildren, FC } from 'react';
+import { FC, ReactNode } from 'react';
 
 export type { FC, ReactElement, ReactNode, ComponentProps } from 'react';
 
-export type FCWithChildren<T = any> = FC<PropsWithChildren & T>;
+export type FCWithChildren<T = unknown> = FC<{ children: ReactNode } & T>;
