@@ -1,11 +1,5 @@
-import { NextPage } from 'next';
-import { AppProps } from 'next/app';
-import { ReactElement, ReactNode } from './components';
+import { FC, FCWithChildren } from './components';
 
-export type NextPageWithLayout<P = object, IP = P> = NextPage<P, IP> & {
-  getLayout?: (page: ReactElement) => ReactNode;
-};
+export type NextPage = FC;
 
-export type AppPropsWithLayout = AppProps & {
-  Component: NextPageWithLayout;
-};
+export type NextLayout = FCWithChildren;
